Menu: Removed ID_WINDOW_SPLIT (Window | Split), since we do not support it anymore.
We could re-enable support if we would get the CSplitterWnd to run.
This would also need lots of testing. A quick check showed that the little
icon for splitting is not shown and splitting itself did not work either.
Anyway, we can live without it.

I left the icon for the command in the resources.